CB-13405 backup and restore should have different controller
