+++ datalake-api/src/main/java/com/sequenceiq/sdx/client/SdxClient.java
@@ -7,8 +7,10 @@ import com.sequenceiq.sdx.api.endpoint.DatabaseServerEndpoint;
 import com.sequenceiq.sdx.api.endpoint.DiagnosticsEndpoint;
 import com.sequenceiq.sdx.api.endpoint.OperationEndpoint;
 import com.sequenceiq.sdx.api.endpoint.ProgressEndpoint;
+import com.sequenceiq.sdx.api.endpoint.SdxBackupEndpoint;
 import com.sequenceiq.sdx.api.endpoint.SdxEndpoint;
 import com.sequenceiq.sdx.api.endpoint.SdxInternalEndpoint;
+import com.sequenceiq.sdx.api.endpoint.SdxRestoreEndpoint;
 import com.sequenceiq.sdx.api.endpoint.SdxUpgradeEndpoint;
 
 public interface SdxClient {
@@ -32,4 +34,8 @@ public interface SdxClient {
     DatabaseServerEndpoint databaseServerEndpoint();
 
     AuthorizationUtilEndpoint authorizationUtilEndpoint();
+
+    SdxBackupEndpoint sdxBackupEndpoint();
+
+    SdxRestoreEndpoint sdxRestoreEndpoint();
 }
